AppBuf: Add the AppBuf class
This class is inspired by the m5 app_buf struct.

This is just the skeleton that will be improved in
next iterations.

Another file, the test_AppBuf.cpp file is also added by this
commit with a basic test for the constructor/destructor.

The Makefile is also updated by this commit.

Signed-off-by: Flavio Santes <flavio.santes@1byt3.com>